+++ src/main/java/duke/Task.java
@@ -1,0 +1,76 @@
+package duke;
+
+public class Task {
+    private String taskDescription;
+    private boolean isDone;
+
+    public Task(String taskDescription) {
+        this.taskDescription = taskDescription;
+        this.isDone = false;
+    }
+
+    public String getStatusIcon() {
+        return (isDone ? "X" : " ");
+    }
+
+    public void setIsDone() {
+        isDone = true;
+    }
+
+    public void revertIsDone() {
+        isDone = false;
+    }
+
+    public String getTaskDes() {
+        return taskDescription;
+    }
+
+    @Override
+    public String toString() {
+        return "[" + this.getStatusIcon() + "]" + " " + this.getTaskDes();
+    }
+}
+
+class Todo extends Task {
+    public Todo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
+class Deadline extends Task {
+    private TimeConvertor by;
+
+    public Deadline(String description, TimeConvertor by) {
+        super(description);
+        this.by = by;
+    }
+
+    public String getDeadline() {
+        return this.by.getDate();
+    }
+
+    @Override
+    public String toString() {
+        return "[D]" + super.toString() + "(by: " + by + ")";
+    }
+}
+
+class Event extends Task {
+    private TimeConvertor from;
+    private TimeConvertor to;
+
+    public Event(String description, TimeConvertor from, TimeConvertor to) {
+        super(description);
+        this.from = from;
+        this.to = to;
+    }
+
+    @Override
+    public String toString() {
+        return "[E]" + super.toString() + "(from: " + from + " to: " + to + ")";
+    }
+}
